datamatrix_barcodes/requirements.txt
pylibdmtx
Pillow
